aggiunta modularizzazione completa
