+++ .gitignore
@@ -1,4 +1,45 @@
+# File di configurazione
 .env
-.cache
-yt-cache/
 data.dat
+
+# Cache e file temporanei
+.cache/
+yt-cache/
+__pycache__/
+*.py[cod]
+*$py.class
+.pytest_cache/
+.coverage
+htmlcov/
+
+# File di log
+*.log
+log.txt
+Error.txt
+
+# File di sistema
+.DS_Store
+Thumbs.db
+desktop.ini
+
+# File di IDE
+.idea/
+.vscode/
+*.swp
+*.swo
+*~
+
+# File di build e distribuzione
+build/
+dist/
+*.egg-info/
+*.egg
+
+# File di ambiente virtuale
+venv/
+env/
+ENV/
+
+# File di output
+downloads/
+*.mp3
