+++ src/__init__.py
@@ -1,0 +1,3 @@
+"""
+SpotifyDl - Un'applicazione per scaricare musica da Spotify
+"""
