+++ src/commands/__init__.py
@@ -1,0 +1,3 @@
+"""
+Modulo per i comandi dell'applicazione
+"""
