+++ src/utils/__init__.py
@@ -1,0 +1,3 @@
+"""
+Modulo per le utility dell'applicazione
+"""
